move pixel_vec module out of `read` to `image`
